# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Metropolitan University Club Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new student/faculty account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /dashboard/club-admin/{slug}:
    get:
      operationId: getClubAdminDashboard
      tags: [dashboard]
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Club admin dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClubAdminDashboard"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]

    # ---------- AUTH ----------
    LoginBody:
      type: object
      properties:
        username: { type: string }
        password: { type: string }
      required: [username, password]
    RegisterBody:
      type: object
      properties:
        username: { type: string }
        password: { type: string }
        fullName: { type: string }
        email: { type: string }
        role:
          type: string
          enum: [student, faculty]
        studentId:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        batch:
          type: ["string", "null"]
      required: [username, password, fullName, email, role]
    User:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        fullName: { type: string }
        email: { type: string }
        role:
          type: string
          enum: [student, faculty, club_admin, overseer]
        studentId:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        batch:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt: { type: string, format: date-time }
      required: [id, username, fullName, email, role, createdAt]
    AuthSession:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        adminClubSlugs:
          type: array
          items: { type: string }
      required: [user, adminClubSlugs]

    # ---------- CLUBS ----------
    Club:
      type: object
      properties:
        id: { type: string }
        slug: { type: string }
        name: { type: string }
        category: { type: string }
        shortDescription: { type: string }
        description: { type: string }
        logoUrl:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        accentColor: { type: string }
        websiteUrl:
          type: ["string", "null"]
        facebookUrl:
          type: ["string", "null"]
        instagramUrl:
          type: ["string", "null"]
        memberCount: { type: integer }
        eventCount: { type: integer }
      required: [id, slug, name, category, shortDescription, description, accentColor, memberCount, eventCount]
...
    Member:
      type: object
      properties:
        id: { type: string }
        userId: { type: string }
        clubId: { type: string }
        clubSlug: { type: string }
        clubName: { type: string }
        fullName: { type: string }
        email: { type: string }
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [president, vice_president, secretary, member]
        joinedAt: { type: string, format: date-time }
      required: [id, userId, clubId, clubSlug, clubName, fullName, email, role, joinedAt]
    JoinRequest:
      type: object
      properties:
        id: { type: string }
        userId: { type: string }
        clubId: { type: string }
        clubSlug: { type: string }
        clubName: { type: string }
        fullName: { type: string }
        email: { type: string }
        studentId:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        batch:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt: { type: string, format: date-time }
      required: [id, userId, clubId, clubSlug, clubName, fullName, email, status, createdAt]
...
    Event:
      type: object
      properties:
        id: { type: string }
        clubId: { type: string }
        clubSlug: { type: string }
        clubName: { type: string }
        title: { type: string }
        description: { type: string }
        startsAt: { type: string, format: date-time }
        endsAt:
          oneOf:
            - { type: string, format: date-time }
            - type: "null"
        venue: { type: string }
        capacity:
          type: ["integer", "null"]
        coverUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        rsvpCount: { type: integer }
        viewerHasRsvp: { type: boolean }
      required: [id, clubId, clubSlug, clubName, title, description, startsAt, venue, status, rsvpCount, viewerHasRsvp]
...
    Post:
      type: object
      properties:
        id: { type: string }
        clubId: { type: string }
        clubSlug: { type: string }
        clubName: { type: string }
        title: { type: string }
        body: { type: string }
        imageUrl:
          type: ["string", "null"]
        authorName: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, clubId, clubSlug, clubName, title, body, authorName, createdAt]
...
    Notice:
      type: object
      properties:
        id: { type: string }
        clubId:
          type: ["string", "null"]
        clubSlug:
          type: ["string", "null"]
        clubName:
          type: ["string", "null"]
        title: { type: string }
        body: { type: string }
        scope:
          type: string
          enum: [club, university]
        pinned: { type: boolean }
        publishAt: { type: string, format: date-time }
        expireAt:
          oneOf:
            - { type: string, format: date-time }
            - type: "null"
        audienceRole:
          type: ["string", "null"]
        createdAt: { type: string, format: date-time }
      required: [id, title, body, scope, pinned, publishAt, createdAt]
...
    ClubAdminDashboard:
      type: object
      properties:
        club:
          $ref: "#/components/schemas/Club"
        members:
          type: array
          items:
            $ref: "#/components/schemas/Member"
        pendingRequests:
          type: array
          items:
            $ref: "#/components/schemas/JoinRequest"
        upcomingEvents:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        pendingEvents:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        recentPosts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        notices:
          type: array
          items:
            $ref: "#/components/schemas/Notice"
      required: [club, members, pendingRequests, upcomingEvents, pendingEvents, recentPosts, notices]
...