# lib/api-zod/src/generated/types/club.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Metropolitan University Club Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Club {
  id: string;
  slug: string;
  name: string;
  category: string;
  shortDescription: string;
  description: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  coverUrl?: string | null;
  accentColor: string;
  /** @nullable */
  websiteUrl?: string | null;
  /** @nullable */
  facebookUrl?: string | null;
  /** @nullable */
  instagramUrl?: string | null;
  memberCount: number;
  eventCount: number;
}