# lib/api-zod/src/generated/types/event.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Metropolitan University Club Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { EventStatus } from "./eventStatus";

export interface Event {
  id: string;
  clubId: string;
  clubSlug: string;
  clubName: string;
  title: string;
  description: string;
  startsAt: Date;
  endsAt?: Date | null;
  venue: string;
  /** @nullable */
  capacity?: number | null;
  /** @nullable */
  coverUrl?: string | null;
  status: EventStatus;
  rsvpCount: number;
  viewerHasRsvp: boolean;
}