# lib/api-zod/src/generated/types/joinRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Metropolitan University Club Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { JoinRequestStatus } from "./joinRequestStatus";

export interface JoinRequest {
  id: string;
  userId: string;
  clubId: string;
  clubSlug: string;
  clubName: string;
  fullName: string;
  email: string;
  /** @nullable */
  studentId?: string | null;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  batch?: string | null;
  /** @nullable */
  message?: string | null;
  status: JoinRequestStatus;
  createdAt: Date;
}