# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Metropolitan University Club Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: string;
  username: string;
  fullName: string;
  email: string;
  role: UserRole;
  /** @nullable */
  studentId?: string | null;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  batch?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: Date;
}